==> src/김홍석/EDA/3_bus_station_preprocess.ipynb <==
# 인근 버스 역 정보 feature 추가

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.notebook import tqdm
import gc

from geopy.distance import geodesic

# 맥에서 글씨 깨짐 방지
from matplotlib import rc
rc('font', family='AppleGothic')
plt.rcParams['axes.unicode_minus'] = False

# Data Load

df = pd.read_csv('../data/preprocessed/2_subway_preprocessed.csv')

df.drop(['Unnamed: 0'], axis=1, inplace=True)

df

bus = pd.read_csv('../data/bus_feature.csv')

bus.head()

# geopy를 사용하여 역과 아파트 사이의 거리 계산
- '좌표X', '좌표Y'를 합쳐서 '좌표X,좌표Y'변수를 생성한다.
    - 모든 데이터를 다 구하면 너무 오래 걸림
    - unique한 값을 구할 수 있다.
    - 연산 횟수를 줄일 수 있다.

- 각각의 좌표에 대해서 역과의 거리를 구한다.
- 다양한 feature를 생성한다.

# '좌표X', '좌표Y'를 합쳐서 하나의 변수 '좌표X,좌표Y'생성

df['좌표X,좌표Y'] = df['좌표X'].astype(str)+','+df['좌표Y'].astype(str)
print(len(df['좌표X,좌표Y'].unique()))

loc_df = pd.DataFrame(df['좌표X,좌표Y'].unique(), columns=['좌표X,좌표Y']) # '좌표X,좌표Y' 부분만 추출
loc_df

# 좌표 X, Y 다시 분리

loc_df[['좌표X', '좌표Y']] = loc_df['좌표X,좌표Y'].apply(lambda x: pd.Series(x.split(',')))

loc_df['좌표X'] = loc_df['좌표X'].astype(float)
loc_df['좌표Y'] = loc_df['좌표Y'].astype(float)

loc_df

# 아파트와 인근 지하철 거리 정보 계산
# 반경 m이하의 정류장 개수
# 가장 가까운 정류장
# cols : 노드 ID	정류소번호	정류소명	X좌표	Y좌표	정류소 타입
m = 300

apart_loc = (loc_df.iloc[0]['좌표Y'], loc_df.iloc[0]['좌표X']) # 아파트 위치 정보
re = bus.apply(lambda loc: geodesic(apart_loc, (loc['Y좌표'], loc['X좌표'])).meters, axis=1) # 아파트와 모든 지하철 정류장 사이의 거리를 구한다.

# 의미있는 정보 표시
print(f'1.5km 이하에 위치한 역 개수: {(re < m).sum()}')
shortest = bus.iloc[re.idxmin()]
print(f'가장 가까운 역과의 거리: {re.min()/1000:.2f}km')
print('가장 가까운 정류장 이름:', shortest['정류소명'])
print('가장 가까운 정류장 타입:', shortest['정류소 타입'])

## 버스 정류장

def calculate_dist(loc_df, bus_df):
    apart_loc = (loc_df['좌표Y'], loc_df['좌표X'])
    distances = bus_df.apply(lambda loc: geodesic(apart_loc, (loc['Y좌표'], loc['X좌표'])).meters, axis=1)
    return distances

distances = loc_df.apply(calculate_dist, bus_df = bus, axis=1)

distances.shape

distances.to_csv('./버스 정류장.csv', index=False)

loc_df['가장 가까운 거리'] = distances.min(axis=1)
loc_df

min_idx = distances.idxmin(axis=1)
bus.iloc[min_idx].reset_index()

loc_df = pd.concat([loc_df, bus.iloc[min_idx].reset_index()], axis=1)
loc_df

(distances<=300).sum(axis=1)

loc_df['인근 버스 정류장 개수'] = (distances<=300).sum(axis=1)

loc_df

### 원본 데이터에 적용

# '시군구 번지'에 맞춰 X, Y 를 매핑

shortest_dist_mapping = pd.Series({locate: shortest_dist for locate, shortest_dist in zip(loc_df['좌표X,좌표Y'], loc_df['가장 가까운 거리'])})
idx_mapping = pd.Series({locate: idx for locate, idx in zip(loc_df['좌표X,좌표Y'], loc_df['index'])})
ID_mapping = pd.Series({locate: ID for locate, ID in zip(loc_df['좌표X,좌표Y'], loc_df['노드 ID'])})
station_num_mapping = pd.Series({locate: station_num for locate, station_num in zip(loc_df['좌표X,좌표Y'], loc_df['정류소번호'])})
station_name_mapping = pd.Series({locate: station_name for locate, station_name in zip(loc_df['좌표X,좌표Y'], loc_df['정류소명'])})
station_type_mapping = pd.Series({locate: station_type for locate, station_type in zip(loc_df['좌표X,좌표Y'], loc_df['정류소 타입'])})
X_mapping = pd.Series({locate: latitude for locate, latitude in zip(loc_df['좌표X,좌표Y'], loc_df['X좌표'])})
Y_mapping = pd.Series({locate: longitude for locate, longitude in zip(loc_df['좌표X,좌표Y'], loc_df['Y좌표'])})
nums_mapping = pd.Series({locate: nums for locate, nums in zip(loc_df['좌표X,좌표Y'], loc_df['인근 버스 정류장 개수'])})

# 새로운 feature 추가
df[['가장 가까운 버스 정류장 거리', '가장 가까운 버스 정류장 index', '가장 가까운 버스 정류장 노드 ID', '가장 가까운 버스 정류소번호', '가장 가까운 버스 정류소명', '가장 가까운 버스 정류소 타입', '가장 가까운 버스 정류장 X좌표', '가장 가까운 버스 정류장 Y좌표', '인근 버스 정류장 개수']] = np.nan

df

# 원본 데이터에 적용
# '가장 가까운 버스 정류장 거리', '가장 가까운 버스 정류장 index'
# '가장 가까운 버스 정류장 노드 ID', '가장 가까운 버스 정류소번호'
# '가장 가까운 버스 정류소명', '가장 가까운 버스 정류소 타입', 
# '가장 가까운 버스 정류장 X좌표', '가장 가까운 버스 정류장 Y좌표', '인근 버스 정류장 개수'

df['가장 가까운 버스 정류장 거리'] = df['가장 가까운 버스 정류장 거리'].fillna(df['좌표X,좌표Y'].map(shortest_dist_mapping))
df['가장 가까운 버스 정류장 index'] = df['가장 가까운 버스 정류장 index'].fillna(df['좌표X,좌표Y'].map(idx_mapping))
df['가장 가까운 버스 정류장 노드 ID'] = df['가장 가까운 버스 정류장 노드 ID'].fillna(df['좌표X,좌표Y'].map(ID_mapping))
df['가장 가까운 버스 정류소번호'] = df['가장 가까운 버스 정류소번호'].fillna(df['좌표X,좌표Y'].map(station_num_mapping))
df['가장 가까운 버스 정류소명'] = df['가장 가까운 버스 정류소명'].fillna(df['좌표X,좌표Y'].map(station_name_mapping))
df['가장 가까운 버스 정류소 타입'] = df['가장 가까운 버스 정류소 타입'].fillna(df['좌표X,좌표Y'].map(station_type_mapping))
df['가장 가까운 버스 정류장 X좌표'] = df['가장 가까운 버스 정류장 X좌표'].fillna(df['좌표X,좌표Y'].map(X_mapping))
df['가장 가까운 버스 정류장 Y좌표'] = df['가장 가까운 버스 정류장 Y좌표'].fillna(df['좌표X,좌표Y'].map(Y_mapping))
df['인근 버스 정류장 개수'] = df['인근 버스 정류장 개수'].fillna(df['좌표X,좌표Y'].map(nums_mapping))

df

df.to_csv('../data/preprocessed/3_bus_preprocessed.csv', index=False)

df[['아파트명', '좌표X', '좌표Y', '가장 가까운 버스 정류소명', '인근 버스 정류장 개수']].iloc[20000:20030]

tmp = pd.read_csv('../data/train_test_data.csv')
tmp.iloc[20000:20030]
