==> src/김홍석/EDA/11_company_EDA.ipynb <==
# k-시공사 데이터 전처리

import pandas as pd
import numpy as np

df = pd.read_csv('../data/preprocessed/10_volume.csv')

df['k-건설사(시공사)'].fillna('unknown', inplace=True)

synonyms = {
    '삼성물산': ['삼성물산주식회사', '삼성물산(주)', '삼성물산 건설부문', '삼성물산(주) 건설부문', '삼성물산,두산건설', '삼성물산(주) 주택부문', "삼성건설", "삼성건설(주)", "삼성래미안", "삼성물산", "(주)삼성물산", "주시회사 현대건설", '삼성물산, 대림산업', '삼성물산, 현대산업개발'],
    '현대건설': ['현대건설(주)', '현대건설주식회사', '(주)현대건설', '현대산업개발', '현대산업개발(주)', 'HDC현대산업개발', '현대bs&c', '현대BS&C', "(주)현대산업개발", "현대산업개발", "주식회사 현대건설", "현대산업개발 주", "현대종합건설", '현대건설,현대산업개발,한진중공업', '현대,대림', '현대 성우건설', '현대,쌍용,GS', '현대건설,삼성물산,현대산업개발', '풍림산업,현대건설'],
    '대림산업': ['대림산업(주)', '대림산업주식회사', '(주)대림산업', '대림건설', '대림건설(주)', "(주)대림", "대림", "대림산업", "대림산업 (주)", '대림,삼성,현대,현산', '대림산업, 롯데건설', '(주)대림/ (주)삼성'],
    '대우건설': ['대우건설(주)', '대우푸르지오', "(주)대우", "(주)대우건설", "대우건설", "주)대우건설", "대우경남", '대우,럭키개발,롯데건설,선경건설,태평양건설', '대우건설,대림산업,태영건설', '남광, 대우', '유원, 선경, 대우건설', '유원, 선경, 대우건설'],
    'GS건설': ['GS건설(주)', '지에스 건설', '지에스건설주식회사', '지에스건설(주)', "(주)gs공사", "GS 건설", "GS건설", "GS건설주식회사", "GS자이 건설", "gs건설", "럭키외"],
    '한라': ['(주)한라', '한라건설', '(주)한라건설', '한라, 두진'],
    '롯데건설': ['롯데건설(주)', '롯데건설주식회사', "(주)롯데건설", "롯데건설", '롯데건설,우성건설'],
    '한화건설': ['(주)한화건설', '한화건설(주)', '삼성물산,한화건설'],
    '두산건설': ['두산건설주식회사', '두산건설(주)', '두산, 이수건설', '두산건설(주)', '두산건설주식회사', '두산, 이수건설', "두산", "두산중공업"],
    '우성건설': ['우성건설(주)', '(주)우성건설', '우성건설(주)'],
    '포스코건설': ['(주)포스코건설', '포스코, 포스코A&C', '포스코', '포스코,SK,현대'],
    'SH공사': ['sh공사', '에스에이치공사', 'SH', "SH공사", "sh 공사", "(주)에스에이치건설", "서울도시개발공사", "주택공사"],
    '건영': ['(주)건영', '(주) 건영', "(주) 건영", "건영", "건영건설", "건영종합건설(주)", " (주) 건영", '(주)건영,삼익주택,라이프주택'],
    '신동아건설': ['신동아건설(주)', '신동아 건설(주)'],
    '계룡건설': ['계룡건설(주)', '계룡건설산업(주)'],
    '코오롱건설': ['코오롱글로벌(주)', '코오롱글로벌', '코오롱건설주식회사'],
    '성원건설': ['성원건설(주)', '(주)성원건설'],
    "금호": ["(주)금호", "(주)금호건설", "(주)금호산업", "금호", "금호건설", "금호건설(주)", "금호산업", "금호자이"],
    "동아건설": ["(주)동아건설산업", "동아건설", "동아건설산업(주)"],
    "동일토건": ["(주)동일토건", "동일토건"],
    "LG건설": ["엘지건설(주)"],

    "삼호건설": ["삼호건설", "삼호주택", "(주)삼호"],
    "서광산업": ["(주)서광건설산업"],
    "서희건설": ["(주)서희건설", "서희건설", "주식회사 서희건설"],
    "신성건설": ["(주)신성건설", "신성건설", "신성건설(주)"],
    "신창건설": ["(주)신창건설", "신창건설", "신창건설(주)->우방건설"],
    "쌍용건설": ["(주)쌍용건설", "쌍용건설", "쌍용건설(주)", '쌍용건설(주)'],
    "우방건설": ["(주)우방", "우방", "우방건설"],
    "우리건설": ["(주)우리건설", "우리건설"],
    "일성건설": ["(주)일성건설", "일성건설", "(주)일성건설"],
    "임광건설": ["임광토건(주)"],
    "청구건설": ["청구", "청구주택"],

    "중앙건설": ["(주)중앙건설", "중앙건설", "중앙건설(주)", "중앙하이츠", '중앙건설(주)'],
    "진흥기업": ["(주)진흥기업", "진흥기업", "진흥기업(주)"],
    "태영건설": ["(주)태영", "(주)태영건설", "태영건설", '태영건설, 현대산업개발'],
    "한양건설": ["(주)한양", "(주)한양건설", "한양", "한양건설", "한양", "한양건설", "한양수자인", "한양주택 (주)", "?양건설", "한양산업개발"],
    "한진중공업": ["(주)한진중공업", "한진중공업", "한진중공업(주)", "한신공영", "한신공영(주)", '한신공영(주), 한양산업개발(주'],
    "한진건설": ["한진건설(주)"],
    "효성건설": ["(주)효성", "(주)효성건설", "효성건설"],
    "KCC건설": ["KCC", "kcc건설"],
    "동부건설": ["동부건설", "동부건설(주)", "동부건설(주)외1"],
    "이수건설": ["이수건설", "이수건설(주)", "이수건설,대우엔지니어링"],
    "삼성중공업": ["삼성중공업", "삼성중공업(주)"],
    "호반건설" : ["(주)호반건설"],
    "LIG건설" : ["LIG건설", "LIG건설(주)"],
    "SK건설": ["(주)에스케이건설"],
    "경남기업": ["경남기업", "경남기업(주)"],
    "경향건설": ["경향건설(주)"],
    "계룡건설": ["계룡건설산업"],
    "고려개발": ["고려개발", "고려개발(주)", "고려산업개발", "고려산업개발(주)"],
    "극동건설" : ["극동건설", "극동건설(주)"],
    "나산건설" : ["나산종합건설"],
    "라이프건설": ["라이프주택"],
    "신도건설": ["(주)신도종합건설", '인덕주택재건축정비사업조합/(주)신도종합건설'],
    "쌍용건설": ["쌍용"],
    "우림건설": ["우림건설", "우림건설(주)"],
    "유원건설": ["유원건설", "유원건설(주)", '울트라건설', '울트라건설(주)', "유원울트라"],
    "월드건설": ['월드건설', '월드건설(주)'],
    '인정건설' : ['인정건설', '인정건설(주)'],
    "흥화건설": ["흥화공업", "흥화공업(주)"],
    "한솔건설": ["한솔"],
    "한보건설": ["한보", "한보철강"],
    "고려개발": ["고려개발 "],
    "동양메이저건설": ["동양메이저(주)"],
    "동일": ["동일토건", "동일하이빌", '동일토건'],
    
    "미주실업": ["미주실업주식회사", '미주실업,동방티앤씨'],
    "벽산건설": ["벽산건설(주)"],
    "삼익건설": ["삼익주택(주)", "삼익주택", "삼익주택(주)", "삼익건설(주)", '삼환, 삼익건설'],
    "삼정건설": ["삼익건설(주)"],
    "삼익주택": ["삼익주택(주)"],
    "삼정건설": ["삼정기업"],
    "삼환": ["삼환기업"],
    "세양건설": ["세양건설산업"],
    "신동아건설": ['신동아'],
    "양우건설": ['양우건설(주)', "양우건설사"],
    "우성건설": ["우성"],
    "우정건설": ["우정건설 주식회사"],
    "진흥건설": ["진흥기업", "진흥연립재건축조합"],
    '풍림건설': ['풍림산업(주)', '풍림산업'],
    '한보건설': ['한보', '한보철상'],
    "(주)신일건업": ['신일산업'],
    '세양건설': ['세양건설산업(주)'],
    '신명건설': ['신명건설,신원건설,미주건설']
}

def replace_synonyms(value, synonyms_dict):
    for key, values in synonyms_dict.items():
        if value in values:
            return key
    return value

df['k-건설사(시공사)'] = df['k-건설사(시공사)'].apply(lambda x: replace_synonyms(x, synonyms))

company_list = sorted(df['k-건설사(시공사)'].unique())
print(len(company_list))

for c in company_list:
    print(c)

construction_companies = {
    1: [
        "GS건설", "KCC건설", "LG건설", "SK건설", "롯데건설", 
        "대림산업", "대우건설", "삼성물산", "포스코건설", "현대건설",
        "한화건설", "두산건설", "쌍용건설", "한진중공업", "코오롱건설", 
        "호반건설", "효성건설", "SH공사", "계룡건설(주)", "계룡건설산업(주)"
    ],
    2: [
        "(주)대보건설", "(주)순영종합건설", "(주)월드건설", "(주)현진종합건설", 
        "BYC", "JR건설", "KR산업", "LIG건설", "TEC건설", "경남기업", 
        "경동건설", "경향건설", "고려개발(주)", "공영토건", "극동건설", 
        "금호", "남광토건(주)", "남해건설", "대명종합건설", "대백건설", 
        "대성산업", "대한주택공사", "대호아이피종합건설", "동부건설", "동아건설", 
        "동양고속건설주식회사", "동원산업건설", "동일토건", "라성건설", "라이프건설", 
        "라인건설", "미원건설(주)", "미주실업", "벽산건설", "보람건설(주)", 
        "삼부토건", "삼익건설", "삼호건설", "삼환", "상부종합건설(주)", 
        "서광산업", "서초건설", "서희건설", "선경건설", "성원건설", 
        "성지건설", "세방기업", "세양건설", "신동아 건설(주)", "신성건설", 
        "신영건설", "신창건설", "아남건설주식회사", "양우건설", "예성종합건설", 
        "우리건설", "우림건설", "우방건설", "우성건설(주)", "이수건설", 
        "일성건설", "임광건설", "자이에스앤디(주)", "중앙건설", "진흥건설", 
        "진흥기업", "청광종합건설(주)", "청구건설", "태영건설", "토왕건설", 
        "풍림건설", "한라", "한보건설", "한양건설", "한일건설"
    ],
    3: [
        "일신건영", "(주) 신한건설", "(주)건설알포메", "(주)대우자동차판매", 
        "(주)도시공감", "(주)삼전건설", "(주)신일건업", "(주)우리창우", 
        "(주)청구주택", "(주)우성건설", "갑을건설", "건영", "고려개발", 
        "고려산업개발", "고속도로관리공단", "금하건설", "나산건설", 
        "대동건설", "대주건설", "대창건설", "대한전척(주)", "도시개발공사", 
        "돈암정릉구역주택재개발정비사업조합", "동선구역주택재개발조합", "동성개발", 
        "동양메이저건설", "동일", "박연정", "문영건설", "삼정건설", 
        "삼중종합건설(주)", "상부종합건설(주)", "선경건설", "세방기업", 
        "신명건설", "신홍선건설", "쌍용건설", "아키데코", "인정건설", 
        "정릉제4구역재개발조합", "주식회사 삼익주택", "청담삼익아파트", 
        "한보철강", "한성강서재건축조합", "한솔건설", "한승", "한일개발", 
        "한진건설", "해주개발주식회사", "해태건설(주)", "현대엔지니어링", 
        "현대힐스테이트", "혜림건설", "홍익종합건업(주)", "효성중공업", 
        "흑석5구역재개발조합", "흥화건설"
    ]
}

len(construction_companies[1])+len(construction_companies[2])+len(construction_companies[3])

# 건설사 랭킹 딕셔너리 생성
ranking_dict = {}
for rank, companies in construction_companies.items():
    for company in companies:
        ranking_dict[company] = rank

df['건설사 랭킹'] = np.nan

df['건설사 랭킹'] = df['k-건설사(시공사)'].map(ranking_dict)

df['건설사 랭킹'].unique()

df['건설사 랭킹'].fillna(2, inplace=True)

df['건설사 랭킹'].unique()

df[['k-건설사(시공사)', '건설사 랭킹']]

df.to_csv('../data/preprocessed/11_const_rank.csv', index=False)
