==> src/김홍석/EDA/14_initial_price_EDA.ipynb <==
# 공시가 데이터 분석

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import gc

# 맥에서 글씨 깨짐 방지
from matplotlib import rc
rc('font', family='AppleGothic')
plt.rcParams['axes.unicode_minus'] = False

# 공시지가 데이터

init_price_li = []

for i in range(2007, 2024):
    init_price_li.append(pd.read_csv(f'../data/added_features/공시가/공시지가_{i}년.csv', encoding='cp949'))

len(init_price_li)

re = []
for i in range(17):
    tmp = pd.DataFrame(init_price_li[i].groupby('시군구명')['공시지가(원/㎡)'].mean()).reset_index()
    tmp.columns = ['시군구명', '공시지가 평균']
    re.append(tmp)

re[0]

# 원본 데이터
- 년도에 따라 나눠서 진행

df = pd.read_csv('../data/preprocessed/13_land_value_per_district.csv')

year_split_df_list = []

for i in range(2007, 2024):
    year_split_df_list.append(df[df['계약년'] == i])

year_split_df_list[0]['계약년'].unique()

year_split_df_list[0]

re[0].columns

# 원본에 적용

def mapping(init_price_df, origin_df):
    init_price_mapping = pd.Series({district: init_price for district, init_price in zip(init_price_df['시군구명'], init_price_df['공시지가 평균'])})

    origin_df['공시지가 평균'] = np.nan
    origin_df['공시지가 평균'] = origin_df['공시지가 평균'].fillna(origin_df['구']).map(init_price_mapping)

    return origin_df

for i in range(17):
    year_split_df_list[i] = mapping(re[i], year_split_df_list[i])

year_split_df_list[0]

df = pd.concat([splited_df for splited_df in year_split_df_list]).sort_index()

df

df.to_csv('../data/preprocessed/14_init_price.csv', index=False)

tmp = pd.read_csv('../data/preprocessed/13_land_value_per_district.csv')

tmp
