==> src/김홍석/EDA/20_ratio_sale_deposit.ipynb <==
# 20. 아파트 매매가격대비 전세가격비율
- iloc[1]

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import math
from sklearn.preprocessing import MinMaxScaler

# 맥에서 글씨 깨짐 방지
from matplotlib import rc
rc('font', family='AppleGothic')
plt.rcParams['axes.unicode_minus'] = False

df = pd.read_excel('../data/added_features/아파트 매매가격대비 전세가격비율_20240715.xlsx')

df

cols = df.columns
cols = [int(str(col)[:10].replace('-', '')[:6]) for col in cols[1:]]

cols.insert(0, '지역명')
cols

df.columns = cols

df.drop(['지역명'], axis=1, inplace=True)
df

df = pd.DataFrame(df.iloc[1]).reset_index()
df.columns = ['년월', '매매 대비 전세가격 비율']

df

# 원본에 적용

origin_df = pd.read_csv('../data/preprocessed/19_rate_trading_price.csv')

origin_df.head(10)

mapping = pd.Series({ym: value for ym, value in zip(df['년월'], df['매매 대비 전세가격 비율'])})

origin_df['매매 대비 전세가격 비율'] = np.nan
origin_df['매매 대비 전세가격 비율'] = origin_df['매매 대비 전세가격 비율'].fillna(origin_df['계약년월']).map(mapping)

origin_df

origin_df['매매 대비 전세가격 비율'].isna().sum()

origin_df.to_csv('../data/preprocessed/20_ratio_sale_deposit.csv', index=False)
