==> src/김홍석/EDA/34_post_price_EDA.ipynb <==
# 34. 이전 3년 가격 정보 추가

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# 맥에서 글씨 깨짐 방지
from matplotlib import rc
rc('font', family='AppleGothic')
plt.rcParams['axes.unicode_minus'] = False

df = pd.read_csv('../data/preprocessed/32_selling_point.csv')

df.head()

# '아파트명'을 기준으로 데이터 분할

apart_list = df['아파트명'].unique().tolist()
apart_list

apart_df_list = []

for apart in apart_list:
    apart_df_list.append(df[df['아파트명']==apart])

# 각 아파트의 2007~2022년도 평균 target

target_mean_list = []

for apart_df in apart_df_list:
    target_mean_list.append(pd.DataFrame(apart_df.groupby('계약년')['target'].mean()))

i=123
print(apart_list[i])
target_mean_list[i]

target_mean_list[0].index[-1]

target_mean_list[0].loc[2023-3:2023, 'target'].mean()

post_target_mean_list = []

for year in range(2010, 2024):
    post_target_mean_list.append(int(target_mean_list[0].loc[year-3:year, 'target'].mean()))


years = [year for year in range(2010, 2024)]

mapping = pd.Series({year:target_mean for year, target_mean in zip(years, post_target_mean_list)})
apart_df_list[0]['최근 3년 평균'] = np.nan
apart_df_list[0]['최근 3년 평균'] = apart_df_list[0]['최근 3년 평균'].fillna(apart_df_list[0]['계약년'].map(mapping))

apart_df_list[0][['계약년월', '최근 3년 평균']]

l = len(apart_list)

for i in range(l):
    post_target_mean_list = []
    years = 
    for year in range(2010, 2024):
        post

mapping = pd.Series({year:target_mean for year, target_mean in zip(years, post_target_mean_list)})
post_target_mean_list[0]['최근 3년 평균'] = np.nan
post_target_mean_list[0]['최근 3년 평균'] = post_target_mean_list[0]['최근 3년 평균'].fillna(post_target_mean_list[0]['계약년'].map(mapping))

post_target_mean_list[0]

# 그냥 각 아파트의 평균 가격 추가

train_df = df[df['is_test']==0]
test_df = df[df['is_test']==1]

apart_target_mean = pd.DataFrame(train_df.groupby('아파트명')['target'].mean()).reset_index()
apart_target_mean

apart_target_mean[apart_target_mean['아파트명']=='개포6차우성']

apart_target_mean_mapping = pd.Series({apart: value for apart, value in zip(apart_target_mean['아파트명'], apart_target_mean['target'])})

df['아파트 평균 가격'] = np.nan
df['아파트 평균 가격'] = df['아파트 평균 가격'].fillna(df['아파트명'].map(apart_target_mean_mapping))

df[['아파트명', '아파트 평균 가격']]

df[df['아파트 평균 가격'].isna()]

## test 데이터에는 존재하지만 train 데이터에는 존재하지 않는 데이터에 대한 결측치 처리
- 11개의 아파트
- 비슷한 성향의 아파트 값으로 대체

- 역삼센트럴2차아이파크: 21000
- 힐데스하임천호: 50700
- 아츠스테이수유점: 14892
- 수유시그니티: 27448
- 더헤리츠: 41000
- 흑석리버파크자이: 65000
- 신수동르끌레브: 47000
- 르엘신반포파크애비뉴: 163000
- 히페리온리버팰리스: 63500
- 세운푸르지오헤리시티: 26600
- 힐스테이트세운센트럴1단지: 23000

df[df['아파트 평균 가격'].isna()]['아파트명'].unique().tolist()

test_df[test_df['아파트명']=='역삼센트럴2차아이파크'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='강남구') & (df['층']==5) & (16<=df['전용면적']) & (df['전용면적']<=20) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='힐데스하임천호'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='강동구') & (df['층']==4) & (80<=df['전용면적']) & (df['전용면적']<=90) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='아츠스테이수유점'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='강북구') & (5<=df['층']) & (df['층']<=15) & (15<=df['전용면적']) & (df['전용면적']<=20) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='수유시그니티'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='강북구') & (10<=df['층']) & (df['층']<=15) & (45<=df['전용면적']) & (df['전용면적']<=50) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='더헤리츠'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='구로구') & (15<=df['층']) & (df['층']<=20) & (20<=df['전용면적']) & (df['전용면적']<=60) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='흑석리버파크자이'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='동작구') & (5<=df['층']) & (df['층']<=10) & (80<=df['전용면적']) & (df['전용면적']<=90) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='신수동르끌레브'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='마포구') & (1<=df['층']) & (df['층']<=5) & (45<=df['전용면적']) & (df['전용면적']<=55) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='르엘신반포파크애비뉴'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='서초구') & (20<=df['층']) & (df['층']<=25) & (55<=df['전용면적']) & (df['전용면적']<=65) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='히페리온리버팰리스'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='성동구') & (1<=df['층']) & (df['층']<=5) & (80<=df['전용면적']) & (df['전용면적']<=90) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='세운푸르지오헤리시티'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='중구') & (10<=df['층']) & (df['층']<=20) & (20<=df['전용면적']) & (df['전용면적']<=30) & (df['is_test']==0)]['아파트 평균 가격'].mean()

test_df[test_df['아파트명']=='힐스테이트세운센트럴1단지'][['구', '층', '전용면적', '좌표Y', '좌표X']]

df[(df['구']=='중구') & (15<=df['층']) & (df['층']<=20) & (20<=df['전용면적']) & (df['전용면적']<=30) & (df['is_test']==0)]['아파트 평균 가격'].mean()

tmp = [21000, 50700, 14892, 27448, 41000, 65000, 47000, 163000, 63500, 26600, 23000]

nan_mapping = pd.Series({apart: value for apart, value in zip(df[df['아파트 평균 가격'].isna()]['아파트명'].unique().tolist(), tmp)})

df['아파트 평균 가격'] = df['아파트 평균 가격'].fillna(df['아파트명'].map(nan_mapping))

df[['아파트명', '아파트 평균 가격']]

df.to_csv('../data/preprocessed/34_apart_price_mean.csv', index=False)

df = pd.read_csv('../data/preprocessed/34_apart_price_mean.csv')
